Rename meta defaults
